# openapi.json
{"openapi":"3.1.0","info":{"title":"Parlance AI-Infused Pipeline","description":"An AI-powered healthcare assistant that helps find doctor contact information using natural language queries. Compatible with Watson Orchestrate for enterprise workflow automation.","contact":{"name":"Parlance AI Team","email":"support@parlance.ai"},"license":{"name":"MIT"},"version":"1.0.0"},"paths":{"/hospital/{hospital_id}":{"get":{"tags":["Hospital Information"],"summary":"Get Hospital Details","description":"Retrieve comprehensive hospital information including departments and doctors by hospital ID","operationId":"get_hospital_details_hospital__hospital_id__get","parameters":[{"name":"hospital_id","in":"path","required":true,"schema":{"type":"integer","title":"Hospital Id"}}],"responses":{"200":{"description":"Hospital details with associated departments and doctors","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Get Hospital Details Hospital  Hospital Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/{user_id}":{"get":{"tags":["User Information"],"summary":"Get User Details","description":"Retrieve user information by user ID","operationId":"get_user_details_user__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Complete user profile information","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Get User Details User  User Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/call_watsonx":{"post":{"tags":["AI Assistant"],"summary":"Find Doctor Phone Number","description":"Use AI to extract doctor phone numbers from natural language queries","operationId":"call_watsonx_call_watsonx_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WatsonxRequest"}}},"required":true},"responses":{"200":{"description":"Doctor's phone number extracted by AI","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DoctorPhoneResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["System"],"summary":"Health Check","description":"Check if the API is running and healthy","operationId":"health_check_health_get","responses":{"200":{"description":"Health status of the API","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Check Health Get"}}}}}}}},"components":{"schemas":{"DoctorPhoneResponse":{"properties":{"response":{"type":"string","title":"Response"}},"type":"object","required":["response"],"title":"DoctorPhoneResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WatsonxRequest":{"properties":{"string_input":{"type":"string","title":"String Input","default":"I would like to get in touch with Dr. Smith from Emergency"},"hospital_id":{"type":"integer","title":"Hospital Id"}},"type":"object","required":["hospital_id"],"title":"WatsonxRequest"}}}}